+++ app/src/js/components/DataUpload/overview.js
@@ -1,5 +1,5 @@
 'use strict';
-import React, { useState, memo } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
@@ -7,24 +7,30 @@ import _config from '../../config';
 import { loadToken } from '../../utils/auth';
 import Loading from '../LoadingIndicator/loading-indicator';
 import localUpload from 'edpub-data-upload-utility';
-import { listFileUploadsBySubmission, listFileDownloadsByKey, refreshToken } from '../../actions';
+import { listFileUploadsBySubmission, listFileDownloadsByKey } from '../../actions';
+import { shortDateShortTimeYearFirstJustValue, storage } from '../../utils/format';
+import Table from '../SortableTable/SortableTable';
 
 class UploadOverview extends React.Component {
   constructor() {
     super();
-    this.state = { loaded: false, hiddenFileInput: React.createRef(null), statusMsg: 'Select a file', uploadFile: '', keys: [] };
+    this.state = { 
+      loaded: false, 
+      hiddenFileInput: React.createRef(null), 
+      statusMsg: 'Select a file', 
+      uploadFile: '', 
+      keys: []
+    };
     this.handleClick = this.handleClick.bind(this);
     this.handleChange = this.handleChange.bind(this);
     this.getFileList = this.getFileList.bind(this);
     this.validateFile = this.validateFile.bind(this);
-    this.resetInputWithTimeout = this.resetInputWithTimeout.bind(this);
     this.keyLookup = this.keyLookup.bind(this);
-    this.isFilePreviouslySaved = this.isFilePreviouslySaved.bind(this);
-    
+    this.resetInputWithTimeout = this.resetInputWithTimeout.bind(this);
   }
-
-  keyLookup(event, fileName) {
-    event.preventDefault();
+  
+  keyLookup(e, fileName) {
+    e.preventDefault();
     if (this.state.keys[fileName]) {
       const { dispatch } = this.props;
       const { requestId } = this.props.match.params;
@@ -34,71 +40,80 @@ class UploadOverview extends React.Component {
         dispatch(listFileDownloadsByKey(this.state.keys[fileName], requestId))
           .then(() => {
             download.downloadFile(this.state.keys[fileName], `${apiRoot}data/upload/downloadUrl`, loadToken().token).then((resp) => {
-              if (resp.error) {
-                console.log(`An error has occured: ${resp.error}.`);
+              let error = resp?.data?.error || resp?.error || resp?.data?.[0]?.error
+              if (error) {
+                console.log(`An error has occurred: ${error}.`);
               }
             })
           }
         );
       }
     }
-  }
-
-  getFileList() {
+  };
+  
+  async getFileList() {
     const { dispatch } = this.props;
     const { requestId } = this.props.match.params;
     if (requestId !== '' && requestId != undefined && requestId !== null) {
       dispatch(listFileUploadsBySubmission(requestId))
         .then((resp) => {
-          let html = [];
-          if (JSON.stringify(resp) === '{}' || (resp.data && resp.data.length === 0)) {
-            this.setState({ saved: 'None found' })
-          } else if (resp.data && resp.data.error) {
-            if (!resp.data.error.match(/not authorized/gi) && !resp.data.error.match(/not implemented/gi)) {
-              const str = `An error has occurred while getting the list of files: ${resp.data.error}.`;
-              this.setState({ saved: str })
+          if (JSON.stringify(resp) === '{}' || JSON.stringify(resp) === '[]' || (resp.data && resp.data.length === 0)) {
+            return
+          }
+          let error = resp?.data?.error || resp?.error || resp?.data?.[0]?.error
+          if (error){
+            if (!error.match(/not authorized/gi) && !error.match(/not implemented/gi)) {
+              const str = `An error has occurred while getting the list of files: ${error}.`;
+              console.log(str)
+              return
             } else {
-              this.setState({ saved: 'None found' })
-            }
-          } else {
-            document.getElementById('previously-saved').replaceChildren();
-            const dataArr = resp.data;
-            if(dataArr.length===0){
-              html.push(<>None found<br /></>)
               return
             }
-            dataArr.sort(function (a, b) {
-              var keyA = new Date(a.last_modified),
-                keyB = new Date(b.last_modified);
-              if (keyA > keyB) return -1;
-              if (keyA < keyB) return 1;
-              return 0;
-            });
-            this.setState({ keys: [] });
-            let tmpKeys = []
-            for (const ea in dataArr) {
-              const fileName = dataArr[ea].file_name;
-              if (dataArr[ea]=== undefined || fileName === undefined){
-                break
-              }
-              const key = dataArr[ea].key;
-              tmpKeys[`${fileName}`] = key
-              if (document.getElementById('previously-saved') !== null) {
-                html.push(<><a id={fileName} name={fileName} aria-label={`Download ${fileName}`} onClick={(e) => this.keyLookup(e, fileName)}>{fileName}</a><br /></>)
-              }
-            }
-            html.map(item =>
-              <span key={item}>{item}</span>
-            )
-            this.setState({ saved: html });
-            this.setState({ keys: tmpKeys });
           }
+
+          const files = resp.data;
+          
+          files.sort(function (a, b) {
+            var keyA = new Date(a.last_modified),
+              keyB = new Date(b.last_modified);
+            if (keyA > keyB) return -1;
+            if (keyA < keyB) return 1;
+            return 0;
+          });
+
+          this.setState({ files: files })
+          
+          const keyDict = {};
+          const html = [];
+
+          for (const ea in files) {
+            const fileName = files[ea].file_name;
+            if (files[ea] === undefined || fileName === undefined) {
+              break
+            }
+            const key = files[ea].key;
+            keyDict[`${fileName}`] = key
+          }
+          this.setState({ files: files })
+
+          html.map(item =>
+            <span key={item}>{item}</span>  
+          )
+
+          this.setState({ keys: keyDict })
+          this.setState({ saved: html })
         });
     }
   }
 
-  componentDidMount() {
-    this.getFileList()
+  async componentDidMount() {
+    const { groupId } = this.props.match.params;
+    const { requestId } = this.props.match.params;
+    if ((requestId !== '' && requestId != undefined && requestId !== null) &&
+      (groupId == '' || groupId === undefined || groupId === null)) {
+      await this.getFileList()
+      this.setState({ loaded: true })
+    }
   }
 
   handleClick(e) {
@@ -107,7 +122,7 @@ class UploadOverview extends React.Component {
     this.state.hiddenFileInput?.current?.click();
   };
 
-  resetInputWithTimeout(msg, timeout){
+  resetInputWithTimeout(msg, timeout) {
     setTimeout(() => {
       msg ? this.setState({ statusMsg: msg }) : null
       if (this.state.hiddenFileInput.current === null || this.state.hiddenFileInput === null) {
@@ -116,35 +131,9 @@ class UploadOverview extends React.Component {
     }, timeout);
   }
 
-  isFilePreviouslySaved(file){
-    let alreadySaved = false;
-    if (this.state.saved) {
-      for (const ea in this.state.saved){
-        let reactElement = this.state.saved[ea]
-        for (const prop in reactElement){
-          if (typeof reactElement[prop] === 'object' && 
-            reactElement[prop] !== null && 
-            reactElement[prop]['children'] !== null && 
-            reactElement[prop]['children']!== undefined && 
-            reactElement[prop]['children'].length > 0){
-            for(const child in reactElement[prop]['children']){
-              if (reactElement[prop]['children'][child]['props']['id'] !== undefined && reactElement[prop]['children'][child]['props']['id'] === file.name) {
-                alreadySaved = true;
-              }
-            }
-          }
-        }
-      }
-    }
-    return alreadySaved
-  }
-
-  validateFile(file){
+  validateFile(file) {
     let valid = false;
-    if (this.isFilePreviouslySaved(file)) {
-      this.setState({ statusMsg: 'This file was already uploaded.' });
-      this.resetInputWithTimeout('Please select a different file.', 2000)
-    } else if (file.name.match(/\.([^\.]+)$/) !== null){
+    if (file.name.match(/\.([^\.]+)$/) !== null) {
       var ext = file.name.match(/\.([^\.]+)$/)[1];
       if (ext.match(/exe/gi)) {
         this.setState({ statusMsg: 'exe is an invalid file type.' });
@@ -158,42 +147,82 @@ class UploadOverview extends React.Component {
     }
     return valid;
   }
-  
+
   async handleChange(e) {
     e.preventDefault();
-    const { dispatch } = this.props;
     const file = e.target.files[0];
-    if(this.validateFile(file)){
+    if (this.validateFile(file)) {
       this.setState({ statusMsg: 'Uploading' });
       const upload = new localUpload();
       const { requestId } = this.props.match.params;
+      const { groupId } = this.props.match.params;
       const { apiRoot } = _config;
-      // TODO - Readd refresh token here. Hitting issue with upload attempt before
-      // token refresh
-      // await dispatch(refreshToken());
-      if (requestId !== '' && requestId != undefined && requestId !== null) {
-        const payload = {
+      try {
+        let payload = {
           fileObj: file,
-          apiEndpoint: `${apiRoot}data/upload/getPostUrl`,
           authToken: loadToken().token,
-          submissionId: requestId
         }
-        await upload.uploadFile(payload).then((resp) => {
-          this.setState({ statusMsg: 'Uploading' });
-          if (resp.error) {
-            console.log(`An error has occured: ${resp.error}.`);
-            this.resetInputWithTimeout('Select a file', 1000)
-          } else {
-            this.setState({ statusMsg: 'Upload Complete' });
-            this.getFileList();
-            this.resetInputWithTimeout('Select another file', 1000)
+        let prefix = ''
+        if (requestId !== '' && requestId != undefined && requestId !== null) {
+          payload['apiEndpoint'] = `${apiRoot}data/upload/getPostUrl`;
+          payload['submissionId'] = requestId
+        } else if (groupId !== '' && groupId != undefined && groupId !== null) {
+          if (document.getElementById('prefix') && document.getElementById('prefix') !== null) {
+            prefix = document.getElementById('prefix').value
           }
-        })
+          payload['apiEndpoint'] = `${apiRoot}data/upload/getGroupUploadUrl`;
+          payload['endpointParams'] = {
+            prefix: prefix,
+            group_id: groupId
+          }
+        }
+        this.setState({ statusMsg: 'Uploading' });
+        const resp = await upload.uploadFile(payload)
+        let error = resp?.data?.error || resp?.error || resp?.data?.[0]?.error
+        if (error) {
+          console.log(`An error has occured on uploadFile: ${error}.`);
+          this.resetInputWithTimeout('Select a file', 1000)
+        } else {
+          this.setState({ statusMsg: 'Upload Complete' });
+          this.resetInputWithTimeout('Select another file', 1000)
+          if ((requestId !== '' && requestId != undefined && requestId !== null) &&
+            (groupId == '' || groupId === undefined || groupId === null)) {
+            this.getFileList()
+          }
+        }
+      } catch (error) {
+        console.log(`try catch error: ${error.stack}`);
+        this.resetInputWithTimeout('Select a file', 1000)
       }
     }
   };
 
   render() {
+    const tableColumns = [
+      {
+        Header: 'Filename',
+        accessor: row => <><a id={row.file_name} name={row.file_name} aria-label={`Download ${row.file_name}`} onClick={(e) => this.keyLookup(e, row.file_name)}> {row.file_name}</a></>,
+        id: 'file_name'
+      },
+      {
+        Header: 'Size',
+        accessor: row => storage(row.size),
+        id: 'size',
+        width: '100px'
+      },
+      {
+        Header: 'sha256Checksum',
+        accessor: row => row.sha256Checksum,
+        id: 'sha256Checksum'
+      },
+      {
+        Header: 'Last Modified',
+        accessor: row => shortDateShortTimeYearFirstJustValue(row.last_modified),
+        id: 'last_modified'
+      }
+    ];
+    const { requestId } = this.props.match.params;
+    const { groupId } = this.props.match.params;
     return (
       <><br></br>
         <div className='page__component'>
@@ -202,10 +231,14 @@ class UploadOverview extends React.Component {
               Upload Data File
             </h1>
           </div>
-          <div className='indented__details'>
+          <div className='indented__details' style={{ paddingTop: '1rem' }}>
             <div className='form__textarea'>
+              {groupId !== undefined ?
+                <><label htmlFor="prefix" style={{ marginBottom: '1rem', marginTop: '1rem', fontSize: 'unset' }}>Subfolder (If applicable): </label><input id="prefix" name="prefix" style={{ marginBottom: '1rem' }} /></>
+              : null
+              }
               {this.state.statusMsg === 'Uploading' ? <Loading /> : null}
-              <label className='heading--medium' htmlFor='hiddenFileInput' style={{ marginBottom: '1rem' }}>{`${this.state.statusMsg}`}
+              <label htmlFor='hiddenFileInput' style={{ marginBottom: '1rem', fontSize: 'unset' }}>{`${this.state.statusMsg}`}
                 <input
                   onChange={(e) => this.handleChange(e)}
                   type="file"
@@ -216,26 +249,36 @@ class UploadOverview extends React.Component {
               </label>
               <button onClick={(e) => this.handleClick(e)} className={'button button--submit button__animation--md button__arrow button__arrow--md button__animation button__arrow--white'}>Upload File</button>
             </div>
-            <br></br><h1 className='heading--small' aria-labelledby='Files Previously Saved'>
-              Files Previously Uploaded:
-            </h1>
-            {!this.state.saved ? <Loading /> : null}
-            <span id='previously-saved'>
-              {this.state.saved ? this.state.saved : null}
-            </span>
+            {this.state.saved && requestId !== undefined && groupId === undefined
+            ?
+              <><br /><section className = 'page__section'>
+                <div style={{ borderBottom: '1px solid #E2DFDF'}}>
+                <h2 className='heading--medium heading--shared-content with-description'>Files Previously Uploaded</h2>
+                </div>
+                <Table
+                  data={this.state.files}
+                  dispatch={this.props.dispatch}
+                  tableColumns={tableColumns}
+                />
+              </section></>
+              : null }
+            {!this.state.loaded && groupId === undefined ? <Loading /> : null}
+            <span>{this.state.saved ? this.state.saved : null}</span>
           </div>
         </div></>
-      );
+    );
   }
 }
 
 UploadOverview.propTypes = {
   stats: PropTypes.object,
   dispatch: PropTypes.func,
-  config: PropTypes.object
+  config: PropTypes.object,
+  logs: PropTypes.object,
 };
 
 export default withRouter(connect(state => ({
   stats: state.stats,
-  config: state.config
+  config: state.config,
+  logs: state.logs
 }))(UploadOverview));