+++ app/src/js/components/Requests/withdrawn.js
@@ -71,6 +71,9 @@ class InactiveRequestsOverview extends React.Component {
       newList[ea] = record;
       for (const r in record) {
         if (record[r].hidden && typeof record[r] === 'object') {
+          if (match === undefined && this.state.filter !== undefined && this.state.filter.length > 0) {
+            match = this.state.filter;
+          }
           const prod = { value: record[r].form_data?.data_producer_info_name, label: record[r].form_data?.data_producer_info_name };
           let dataProduct = record[r].form_data?.data_product_name_value;
           if (dataProduct === undefined) {